+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Types of changes
 ## [3.6.0]
 
 - `Added` new push mode `upsert` to insert or update rows based on primary keys
+- `Fixed` relations extraction with Oracle database create duplicated relations
 
 ## [3.5.0]